Add Aido . Change gateway-machine relation
